Changes to be committed: 	modified:   NEWS 	modified:   bin/breadbot 	modified:   breadbot/__init__.py 	modified:   breadbot/core/__init__.py 	modified:   breadbot/core/data.py 	modified:   breadbot/core/dia.py 	modified:   breadbot/core/klg.py 	modified:   breadbot/core/memo.py 	modified:   breadbot/core/misc.py 	modified:   breadbot/core/teach.py 	deleted:    breadbot/log/__init__.py 	deleted:    breadbot/log/mem.log 	modified:   breadbot/serv/wechat/views.py 	modified:   data/dia/knowledge/computers.yml 	modified:   data/klg/computer.yml 	modified:   data/new.yml 	modified:   etc/bread.cfg 	modified:   setup.cfg 	modified:   setup.py

+++ breadbot/core/data.py
@@ -32,7 +32,8 @@ class Data(object):
         data_log = os.path.join(
             misc.cfg().get('log_path'),
             'data.log')
-        os.remove(data_log)
+        if os.path.exists(data_log):
+            os.remove(data_log)
         print('\n Drop database done.')
 
     def _open_db(self):
